fix config value to float

+++ internal/fta/plugins/pftool/pftool.go
@@ -18,8 +18,8 @@ const (
 var _ plugin.ConduitFTAPlugin = (*PftoolPlugin)(nil)
 
 type ViperPftoolPluginConfig struct {
-	PfcpPath     string `mapstructure:"pfcp-path" yaml:"pfcp-path"`
-	TimeoutHours int    `mapstructure:"no-progress-timeout-hours" yaml:"no-progress-timeout-hours"`
+	PfcpPath     string  `mapstructure:"pfcp-path" yaml:"pfcp-path"`
+	TimeoutHours float64 `mapstructure:"no-progress-timeout-hours" yaml:"no-progress-timeout-hours"`
 }
 
 type PftoolPlugin struct {
